+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.16.6
 
 * trace_exception can take None as an argument
+* added option to specify a custom metric source for